+++ client/plots/matrix.js
@@ -10,7 +10,7 @@ export { getPlotConfig } from './matrix.config'
 import { getLegendData, getLegendItemText } from './matrix.legend'
 import { getSerieses } from './matrix.serieses'
 import { setAutoDimensions, setLabelsAndScales, setLayout } from './matrix.layout'
-import { computeStateDiff, mayRequireToken, setData } from './matrix.data'
+import { computeStateDiff, mayRequireToken, setData, applyLegendValueFilter } from './matrix.data'
 import {
 	getTermOrder,
 	getSampleGroups,
@@ -175,12 +175,17 @@ export class Matrix {
 			// may skip term or sample ordering when there are
 			// no relevant state/config/setting changes
 			if (this.stateDiff.nonsettings || this.stateDiff.sorting) {
+				applyLegendValueFilter(this)
 				this.termOrder = this.getTermOrder(this.data)
 				this.sampleGroups = this.getSampleGroups(this.hierClusterSamples || this.data)
 				this.sampleOrder = this.getSampleOrder(this.data)
 			}
 
-			if (!this.sampleOrder.length) {
+			if (
+				!this.sampleOrder.length &&
+				!this.config.legendGrpFilter.lst.length &&
+				!this.config.legendValueFilter.lst.length
+			) {
 				this.dom.loadingDiv.html('No matching sample data').style('display', '')
 				this.dom.svg.style('display', 'none')
 				return
@@ -280,6 +285,7 @@ Matrix.prototype.setLayout = setLayout
 Matrix.prototype.computeStateDiff = computeStateDiff
 Matrix.prototype.mayRequireToken = mayRequireToken
 Matrix.prototype.setData = setData
+Matrix.prototype.applyLegendValueFilter = applyLegendValueFilter
 
 // from matrix.groups
 Matrix.prototype.getTermOrder = getTermOrder