+++ server/src/termdb.scatter.js
@@ -190,7 +190,7 @@ async function colorAndShapeSamples(refSamples, cohortSamples, data, q) {
 	for (const sample of cohortSamples) {
 		const dbSample = data.samples[sample.sampleId.toString()]
 		if (!dbSample) {
-			console.log(JSON.stringify(sample) + ' not in the database or filtered')
+			//console.log(JSON.stringify(sample) + ' not in the database or filtered')
 			continue
 		}
 		let isLast = false
@@ -205,7 +205,9 @@ async function colorAndShapeSamples(refSamples, cohortSamples, data, q) {
 					isLast = true
 				}
 			} else {
-				const key = dbSample[q.divideByTW.term.id || q.divideByTW.term.name]?.key
+				const field = q.divideByTW.term.id || q.divideByTW.term.name
+				const key = dbSample[field]?.key
+				if (key == null) continue
 				divideBy = q.divideByTW.term.values?.[key]?.label || key
 			}
 		}